+++ package.json
@@ -2,25 +2,13 @@
     "name": "sprotty-view-filtering",
     "description": "Please enter a brief description here",
     "scripts": {
-        "build:server": "tsc",
-        "build:browser": "esbuild ./src/diagram/index.ts --bundle --sourcemap --outfile=./static/index.js",
-        "build": "npm run build:server && npm run build:browser",
-        "watch:server": "tsc -w",
-        "watch:browser": "npm run build:browser -- --watch",
-        "start": "node ./lib/server/server"
+        "build": "npm run build --workspaces",
+        "start": "node ./packages/app/lib/server/server"
 
     },
-    "devDependencies": {
-        "@types/node": "^18.0.0",
-        "@types/ws": "^8.5.4",
-        "esbuild": "^0.17.8",
-        "express": "^4.18.2",
-        "typescript": "4.9.5",
-        "ws": "^8.12.1"
-    },
-    "dependencies": {
-        "reflect-metadata": "^0.1.13",
-        "sprotty": "next",
-        "sprotty-elk": "next"
-    }
+    "workspaces": [
+        "packages/common",
+        "packages/sprotty-views",
+        "packages/app"
+    ]
 }
