+++ packages/app/src/server/data/ai-papers.json
