+++ packages/common/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "common",
+    "version": "1.0.0",
+    "description": "Types, Sprotty Actions used across packages",
+    "main": "lib/index",
+    "types": "lib/index",
+    "scripts": {
+        "build": "tsc",
+        "watch": "tsc -w"
+    },
+    "devDependencies": {
+        "@types/node": "^18.0.0",
+        "typescript": "4.9.5"
+    },
+    "dependencies": {
+        "sprotty-protocol": "next"
+    }
+}
