+++ packages/common/src/data.ts
@@ -1,0 +1,9 @@
+export interface OptimizeData {
+    useIsVisible: boolean;
+    useZoomFactor: boolean;
+}
+
+export const optimizeData: OptimizeData = {
+    useIsVisible: false,
+    useZoomFactor: false
+}
