+++ scripts/generateData.ts
@@ -39,5 +39,5 @@ const maxEdgesPerNode = 8; // Change as needed
 const graph = generateRandomGraph(numNodes, maxEdgesPerNode);
 const jsonGraph = JSON.stringify(graph, null, 2);
 
-fs.writeFileSync('../src/server/data/random_graph.json', jsonGraph);
+fs.writeFileSync('../packages/app/src/server/data/random_graph.json', jsonGraph);
 console.log('Random graph data saved to random_graph.json');
