--- tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "commonjs",
-    "types": [
-      "reflect-metadata",
-      "node"
-    ],
-    "experimentalDecorators": true,
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "sourceMap": true,
-    "jsx": "react",
-    "outDir": "./lib"
-  },
-  "include": [
-    "./src/server/**/*.ts",
-    "./src/common/**/*.ts"
-  ],
-  "lib": [
-    "ES2022",
-    "DOM"
-  ]
-}
